pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finm"
version = "0.1.0"
description = "A student-led Python package for financial mathematics and quantitative finance education"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "University of Chicago Financial Mathematics Program" },
]
keywords = [
    "finance",
    "financial-mathematics",
    "quantitative-finance",
    "fixed-income",
    "education",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering :: Mathematics",
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "matplotlib>=3.5.0",
    "scipy>=1.8.0",
    "QuantLib",
]

[project.scripts]
finm = "finm.data._cli:main"

[project.optional-dependencies]
data = [
    "pandas-datareader>=0.10.0",
]
cli = [
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pandas-datareader>=0.10.0",
]
docs = [
    "sphinx>=7.3.7",
    "sphinx-book-theme>=1.1.3",
    "sphinx-autodoc2>=0.5.0",
    "myst-nb>=1.1.2",
    "myst-parser>=2.0.0",
    "sphinx_design>=0.6.1",
    "sphinx-copybutton>=0.5.2",
    "linkify-it-py>=2.0.3",
]
build = [
    "doit>=0.36.0",
    "jupytext>=1.16.0",
    "jupyter>=1.0.0",
    "nbconvert>=7.0.0",
    "colorama>=0.4.6",
    "python-dotenv>=1.0.0",
]
all = [
    "finm[data,cli,dev,docs,build]",
]

[project.urls]
Homepage = "https://github.com/uchicago-finmath/finm"
Documentation = "https://github.com/uchicago-finmath/finm#readme"
Repository = "https://github.com/uchicago-finmath/finm"

[tool.hatch.build.targets.wheel]
packages = ["src/finm"]

[tool.hatch.build]
include = [
    "src/finm/**/*.csv",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]
exclude = [
    "/src/*.py",
    "/src/*.ipynb",
]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "pytest --cov=finm --cov-report=term-missing {args:tests}"

[tool.hatch.envs.lint]
dependencies = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.hatch.envs.lint.scripts]
fmt = "black src tests"
check = "ruff check src tests"
typing = "mypy src"
all = ["fmt", "check", "typing"]

[tool.black]
target-version = ["py39"]
line-length = 88

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    # Ignore QuantLib SWIG deprecation warnings (external library)
    "ignore:builtin type Swig.*:DeprecationWarning",
]

[tool.hatch.envs.docs]
features = ["docs"]

[tool.hatch.envs.docs.scripts]
build = "sphinx-build -b html docs_src docs"

